refactor(postcss-minify-selectors): split foldToIs helpers into a sibling module
