+++ packages/postcss-minify-selectors/src/lib/foldToIs.js
@@ -1,236 +1,14 @@
 'use strict';
-
-// Folds sibling selectors in a comma-separated list by factoring out a common
-// prefix and/or suffix into a shared `:is()` group.
-//
-// The specificity guard below is what keeps the rewrite safe w.r.t. the
-// cascade: `:is(a, b)` takes the max specificity of its arguments, so folding
-// selectors whose divergent parts have different specificities would silently
-// change which rule wins at match time.
-
-/**
- * @typedef {object} Token
- * @property {'compound'|'combinator'} kind
- * @property {string} str
- * @property {import('postcss-selector-parser').Node[]} [nodes]
- */
+const {
+  tokenize,
+  hasPseudoElementOrNesting,
+  hasNthChildOfClause,
+  specificityOfMiddle,
+  sameSpecificity,
+  joinTokens,
+} = require('./foldToIsHelpers.js');
 
 /**
- * @param {import('postcss-selector-parser').Selector} selector
- * @return {Token[]}
- */
-function tokenize(selector) {
-  /** @type {Token[]} */
-  const tokens = [];
-  /** @type {import('postcss-selector-parser').Node[]} */
-  let bucket = [];
-
-  const flush = () => {
-    if (bucket.length) {
-      tokens.push({
-        kind: 'compound',
-        str: bucket.map((n) => String(n)).join(''),
-        nodes: bucket,
-      });
-      bucket = [];
-    }
-  };
-
-  for (const node of selector.nodes) {
-    if (node.type === 'combinator') {
-      flush();
-      tokens.push({ kind: 'combinator', str: String(node) });
-    } else {
-      bucket.push(node);
-    }
-  }
-  flush();
-  return tokens;
-}
-
-/**
- * Pseudo-elements cannot live inside `:is()`; nesting `&` would change meaning.
- *
- * @param {Token} token
- * @return {boolean}
- */
-function hasPseudoElementOrNesting(token) {
-  if (token.kind !== 'compound' || !token.nodes) {
-    return false;
-  }
-  for (const n of token.nodes) {
-    if (n.type === 'nesting') {
-      return true;
-    }
-    if (n.type === 'pseudo') {
-      const v = n.value;
-      if (v.startsWith('::')) {
-        return true;
-      }
-      if (
-        v === ':before' ||
-        v === ':after' ||
-        v === ':first-letter' ||
-        v === ':first-line'
-      ) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-/**
- * `:nth-child(An+B of S)` / `:nth-last-child(An+B of S)` add the specificity
- * of `S` to the pseudo-class. The selector parser flattens `An+B` and the
- * `of S` clause into a single child selector, so we can't cleanly recover S's
- * spec. Reject folding any compound that uses this syntax to stay safe.
- *
- * @param {Token} token
- * @return {boolean}
- */
-function hasNthChildOfClause(token) {
-  if (token.kind !== 'compound' || !token.nodes) {
-    return false;
-  }
-  for (const n of token.nodes) {
-    if (n.type !== 'pseudo') {
-      continue;
-    }
-    if (n.value !== ':nth-child' && n.value !== ':nth-last-child') {
-      continue;
-    }
-    for (const child of n.nodes) {
-      for (const inner of child.nodes) {
-        if (inner.type === 'tag' && inner.value === 'of') {
-          return true;
-        }
-      }
-    }
-  }
-  return false;
-}
-
-/** @typedef {[number, number, number]} Specificity */
-
-/**
- * Specificity contribution of a list of simple-selector nodes, following the
- * nested-specificity rules for `:is()`, `:not()`, `:has()`, and `:where()`.
- *
- * @param {import('postcss-selector-parser').Node[]} nodes
- * @return {Specificity}
- */
-function specificityOf(nodes) {
-  let id = 0;
-  let cls = 0;
-  let type = 0;
-  for (const n of nodes) {
-    if (n.type === 'id') {
-      id++;
-    } else if (n.type === 'class' || n.type === 'attribute') {
-      cls++;
-    } else if (n.type === 'pseudo') {
-      const v = n.value;
-      if (v.startsWith('::')) {
-        type++;
-        continue;
-      }
-      if (v === ':where') {
-        continue;
-      }
-      if (v === ':is' || v === ':matches' || v === ':not' || v === ':has') {
-        const s = maxChildSpecificity(n);
-        id += s[0];
-        cls += s[1];
-        type += s[2];
-        continue;
-      }
-      // Legacy single-colon pseudo-elements count as type.
-      if (
-        v === ':before' ||
-        v === ':after' ||
-        v === ':first-letter' ||
-        v === ':first-line'
-      ) {
-        type++;
-      } else {
-        cls++;
-      }
-    } else if (n.type === 'tag') {
-      type++;
-    }
-  }
-  return [id, cls, type];
-}
-
-/**
- * @param {import('postcss-selector-parser').Pseudo} pseudo
- * @return {Specificity}
- */
-function maxChildSpecificity(pseudo) {
-  /** @type {Specificity} */
-  let best = [0, 0, 0];
-  for (const child of pseudo.nodes) {
-    if (child.type !== 'selector') {
-      continue;
-    }
-    const s = specificityOf(child.nodes);
-    if (compareSpecificity(s, best) > 0) {
-      best = s;
-    }
-  }
-  return best;
-}
-
-/**
- * @param {Token[]} middle
- * @return {Specificity}
- */
-function specificityOfMiddle(middle) {
-  let id = 0;
-  let cls = 0;
-  let type = 0;
-  for (const token of middle) {
-    if (token.kind !== 'compound' || !token.nodes) {
-      continue;
-    }
-    const s = specificityOf(token.nodes);
-    id += s[0];
-    cls += s[1];
-    type += s[2];
-  }
-  return [id, cls, type];
-}
-
-/**
- * @param {Specificity} a
- * @param {Specificity} b
- * @return {number}
- */
-function compareSpecificity(a, b) {
-  return a[0] - b[0] || a[1] - b[1] || a[2] - b[2];
-}
-
-/**
- * @param {Specificity} a
- * @param {Specificity} b
- * @return {boolean}
- */
-function sameSpecificity(a, b) {
-  return a[0] === b[0] && a[1] === b[1] && a[2] === b[2];
-}
-
-/**
- * @param {Token[]} tokens
- * @return {string}
- */
-function joinTokens(tokens) {
-  return tokens.map((t) => t.str).join('');
-}
-
-/**
- * Returns the folded selector list string if beneficial, otherwise null.
- *
  * @param {import('postcss-selector-parser').Root} root
  * @return {string | null}
  */
@@ -275,9 +53,6 @@ function tryFold(root) {
     suffix++;
   }
 
-  // Each prefix/suffix boundary must land on a combinator — otherwise
-  // splicing `:is(...)` in would place it adjacent to a type/class token
-  // in the prefix/suffix and silently change what the selector matches.
   const firstTokens = tokenLists[0];
   while (prefix > 0 && firstTokens[prefix - 1].kind !== 'combinator') {
     prefix--;
@@ -339,14 +114,4 @@ function tryFold(root) {
   return folded;
 }
 
-module.exports = {
-  tryFold,
-  tokenize,
-  hasPseudoElementOrNesting,
-  hasNthChildOfClause,
-  specificityOf,
-  specificityOfMiddle,
-  maxChildSpecificity,
-  compareSpecificity,
-  sameSpecificity,
-};
+module.exports = tryFold;
